+++ src/Request/HandleValueRequest.mjs
@@ -1,6 +1,6 @@
 import { HttpServerResponse } from "../../../flux-http-api/src/Server/HttpServerResponse.mjs";
 import { ILIAS_OBJECT_ID_PATTERN } from "../Ilias/ILIAS_OBJECT_ID.mjs";
-import { METHOD_DELETE, METHOD_GET, METHOD_HEAD, METHOD_PATCH, METHOD_PUT } from "../../../flux-http-api/src/Method/METHOD.mjs";
+import { METHOD_DELETE, METHOD_GET, METHOD_HEAD, METHOD_PATCH, METHOD_POST, METHOD_PUT } from "../../../flux-http-api/src/Method/METHOD.mjs";
 import { STATUS_CODE_400, STATUS_CODE_404 } from "../../../flux-http-api/src/Status/STATUS_CODE.mjs";
 
 /** @typedef {import("../FluxFieldValueStorage/FluxFieldValueStorageService.mjs").FluxFieldValueStorageService} FluxFieldValueStorageService */
@@ -359,7 +359,8 @@ export class HandleValueRequest {
             request,
             [
                 METHOD_GET,
-                METHOD_HEAD
+                METHOD_HEAD,
+                METHOD_POST
             ]
         );
 
@@ -367,8 +368,24 @@ export class HandleValueRequest {
             return response;
         }
 
+        let filter;
+        if (request.method === METHOD_POST) {
+            try {
+                filter = await request.body.json();
+            } catch (error) {
+                console.error(error);
+
+                return HttpServerResponse.text(
+                    "Invalid filter",
+                    STATUS_CODE_400
+                );
+            }
+        } else {
+            filter = Object.fromEntries(request.url.searchParams);
+        }
+
         const values = await this.#flux_field_value_storage_service.getValues(
-            Object.fromEntries(request.url.searchParams)
+            filter
         );
 
         if (values === null) {
@@ -392,7 +409,8 @@ export class HandleValueRequest {
             request,
             [
                 METHOD_GET,
-                METHOD_HEAD
+                METHOD_HEAD,
+                METHOD_POST
             ]
         );
 
@@ -400,8 +418,24 @@ export class HandleValueRequest {
             return response;
         }
 
+        let filter;
+        if (request.method === METHOD_POST) {
+            try {
+                filter = await request.body.json();
+            } catch (error) {
+                console.error(error);
+
+                return HttpServerResponse.text(
+                    "Invalid filter",
+                    STATUS_CODE_400
+                );
+            }
+        } else {
+            filter = Object.fromEntries(request.url.searchParams);
+        }
+
         const table = await this.#flux_field_value_storage_service.getValueTable(
-            Object.fromEntries(request.url.searchParams)
+            filter
         );
 
         if (table === null) {